app/requirements.txt
# Web framework
fastapi~=0.115
uvicorn[standard]~=0.30  # ASGI server for FastAPI

# Database / ORM
sqlalchemy~=2.0
pydantic~=2.9             # FastAPI models & validation
python-dotenv~=1.0        # Load env vars from .env

# Numerical / scientific
numpy~=2.1.2              # kept from original for simulation math

# Testing
pytest~=8.3
httpx~=0.27               # for async API testing if needed

# Dev tools
black~=24.8.0
isort~=5.13
mypy~=1.11
